requirements.txt
# Basic dependencies
fastapi~=0.115.11
uvicorn~=0.34.0
SQLAlchemy~=2.0.39
pydantic~=2.10.6
python-dotenv~=1.0.1

# AI Integration
google-generativeai~=0.1.0
protobuf~=5.29.3
google-api-python-client~=2.164.0
google-auth-httplib2~=0.2.0
httplib2~=0.22.0

# Analytics
nltk~=3.9.1

# Testing
pytest~=8.3.5
pytest-mock~=3.14.0
pytest-cov~=4.1.0
httpx~=0.28.1